chore: move `ignore` related logic into one place
